Migrate

+++ database/migrations/2020_06_07_122059_create_technology_masters_table.php
@@ -17,7 +17,7 @@ class CreateTechnologyMastersTable extends Migration
             $table->id();
             $table->string('code');
             $table->string('name');
-            $table->integer('category_id')->unsigned();
+            $table->bigInteger('category_id')->unsigned();
             $table->foreign('category_id')
             ->references('id')->on('category_masters')
             ->onDelete('cascade');
