+++ database/migrations/2020_06_07_123032_create_user_tech_relates_table.php
@@ -15,12 +15,12 @@ class CreateUserTechRelatesTable extends Migration
     {
         Schema::create('user_tech_relates', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_info_id')->unsigned();
+            $table->bigInteger('user_info_id')->unsigned();
             $table->foreign('user_info_id')
             ->references('id')->on('user_infoes')
             ->onDelete('cascade');
 
-            $table->integer('technology_master_id')->unsigned();
+            $table->bigInteger('technology_master_id')->unsigned();
             $table->foreign('technology_master_id')
             ->references('id')->on('technology_masters')
             ->onDelete('cascade');
